+++ crypto-onramp/src/main/java/com/stripe/android/crypto/onramp/model/OnrampCallbacks.kt
@@ -20,6 +20,7 @@ class OnrampCallbacks {
     private var authorizeCallback: OnrampAuthorizeCallback? = null
     private var checkoutCallback: OnrampCheckoutCallback? = null
     private var userAttestationCallback: OnrampUserAttestationCallback? = null
+    private var termsAndConditionsCallback: OnrampTermsAndConditionsCallback? = null
     private var onrampSessionClientSecretProvider: OnrampSessionClientSecretProvider? = null
     private var googlePayIsReadyCallback: ((Boolean) -> Unit)? = null
     private var samsungPayIsReadyCallback: ((Boolean, SamsungPayAvailabilityResult) -> Unit)? = null
@@ -67,6 +68,13 @@ class OnrampCallbacks {
     }
 
     /**
+     * Callback invoked when terms and conditions presentation completes.
+     */
+    fun termsAndConditionsCallback(callback: OnrampTermsAndConditionsCallback) = apply {
+        this.termsAndConditionsCallback = callback
+    }
+
+    /**
      * An async closure that calls your backend to perform a checkout.
      *     Your backend should call Stripe's `/v1/crypto/onramp_sessions/:id/checkout` endpoint with the session ID.
      *     The closure should return the onramp session client secret on success, or throw an Error on failure.
@@ -106,6 +114,7 @@ class OnrampCallbacks {
         val authorizeCallback: OnrampAuthorizeCallback,
         val checkoutCallback: OnrampCheckoutCallback,
         val userAttestationCallback: OnrampUserAttestationCallback?,
+        val termsAndConditionsCallback: OnrampTermsAndConditionsCallback?,
         val onrampSessionClientSecretProvider: OnrampSessionClientSecretProvider,
         val googlePayIsReadyCallback: ((Boolean) -> Unit)?,
         val samsungPayIsReadyCallback: ((Boolean, SamsungPayAvailabilityResult) -> Unit)?,
@@ -129,6 +138,7 @@ class OnrampCallbacks {
                 "checkoutCallback must not be null"
             },
             userAttestationCallback = userAttestationCallback,
+            termsAndConditionsCallback = termsAndConditionsCallback,
             onrampSessionClientSecretProvider = requireNotNull(onrampSessionClientSecretProvider) {
                 "onrampSessionClientSecretProvider must be not null"
             },